+++ R/form_schema.R
@@ -31,6 +31,12 @@
 #'   Default: TRUE.
 #' @param draft Whether the form is published (FALSE) or a draft (TRUE).
 #'   Default: TRUE.
+#' @param version (character) The published Form version whose schema
+#'   fields to return, e.g. from `form_version_list()`.
+#'   Pass `___` for a blank version.
+#'   If given, the version path is used instead of the published or
+#'   draft path.
+#'   Default: `NULL`.
 #' @template param-pid
 #' @template param-fid
 #' @template param-url
@@ -64,6 +70,7 @@
 #'   }
 # nolint start
 #' @seealso \url{https://docs.getodk.org/central-api-form-management/#getting-form-schema-fields}
+#' @seealso \url{https://docs.getodk.org/central-api-form-management/#getting-form-version-schema-fields}
 # nolint end
 #' @family form-management
 #' @export
@@ -149,18 +156,21 @@
 #' # Point location: used by handle_ru_geopoints
 #' fs %>% dplyr::filter(type == "geopoint")
 #' }
-form_schema <- function(flatten = FALSE,
-                        odata = FALSE,
-                        parse = TRUE,
-                        draft = FALSE,
-                        pid = get_default_pid(),
-                        fid = get_default_fid(),
-                        url = get_default_url(),
-                        un = get_default_un(),
-                        pw = get_default_pw(),
-                        odkc_version = get_default_odkc_version(),
-                        retries = get_retries(),
-                        verbose = get_ru_verbose()) {
+form_schema <- function(
+  flatten = FALSE,
+  odata = FALSE,
+  parse = TRUE,
+  draft = FALSE,
+  version = NULL,
+  pid = get_default_pid(),
+  fid = get_default_fid(),
+  url = get_default_url(),
+  un = get_default_un(),
+  pw = get_default_pw(),
+  odkc_version = get_default_odkc_version(),
+  retries = get_retries(),
+  verbose = get_ru_verbose()
+) {
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
   ru_msg_info(glue::glue("Form schema v{odkc_version}"), verbose = verbose)
 
@@ -199,7 +209,20 @@ form_schema <- function(flatten = FALSE,
     return(fs)
   } else {
     # nocov end
-    if (draft == FALSE) {
+    if (!is.null(version)) {
+      if (
+        !is.character(version) ||
+          length(version) != 1L ||
+          is.na(version) ||
+          !nzchar(version)
+      ) {
+        ru_msg_abort("version must be a single non-empty character string.")
+      }
+      pth <- glue::glue(
+        "v1/projects/{pid}/forms/{URLencode(fid, reserved = TRUE)}/",
+        "versions/{URLencode(version, reserved = TRUE)}/fields"
+      )
+    } else if (draft == FALSE) {
       pth <- glue::glue(
         "v1/projects/{pid}/forms/{URLencode(fid, reserved = TRUE)}/fields"
       )
@@ -221,7 +244,8 @@ form_schema <- function(flatten = FALSE,
       httr::content(.) %>%
       tibble::tibble(xx = .) %>%
       tidyr::unnest_wider(xx) %>%
-      { # nolint
+      {
+        # nolint
         if ("path" %in% names(.)) {
           dplyr::mutate(
             .,